PDV_Extensionista-master/requirements.txt
customtkinter
mysql-connector-python
requests
matplotlib
# Se você tiver mais alguma biblioteca instalada que não está aqui, adicione.